Organise class into packages and fix formatting issues
