+++ src/main/java/bestie/Task.java
@@ -1,3 +1,4 @@
+package bestie;
 abstract public class Task {
 
     protected String description;
