+++ src/main/java/bestie/command/AddCommand.java
@@ -1,4 +1,10 @@
-public class AddCommand extends Command{
+package bestie.command;
+
+import bestie.Storage;
+import bestie.TaskList;
+import bestie.Ui;
+import bestie.Task;
+public class AddCommand extends Command {
 
     private Task task;
     public AddCommand(Task task) {
