+++ src/main/java/bestie/command/EmptyCommand.java
@@ -1,3 +1,9 @@
+package bestie.command;
+
+import bestie.Storage;
+import bestie.TaskList;
+import bestie.Ui;
+
 public class EmptyCommand extends Command {
 
     @Override
