+++ src/main/java/bestie/command/ListCommand.java
@@ -1,4 +1,10 @@
-public class ListCommand extends Command{
+package bestie.command;
+
+import bestie.Storage;
+import bestie.TaskList;
+import bestie.Ui;
+
+public class ListCommand extends Command {
     // display all the tasks in the list
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
