+++ src/main/java/bestie/command/MarkCommand.java
@@ -1,16 +1,23 @@
-public class UnmarkCommand extends Command {
+package bestie.command;
+
+import bestie.Storage;
+import bestie.TaskList;
+import bestie.Ui;
+
+public class MarkCommand extends Command {
 
     private int index; // index of task that has been marked
 
-    public UnmarkCommand(int index) {
+    public MarkCommand(int index) {
         this.index = index;
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
         if (this.index >= 0 && this.index < tasks.size()) {
-            tasks.getTask(this.index).markUndone();
-            ui.showTaskUnmarked(tasks.getTask(this.index));
+
+            tasks.getTask(this.index).markTaskDone();
+            ui.showTaskMarked(tasks.getTask(this.index));
         } else {
             ui.showIndexOutOfBoundsMessage(this.index);
         }
