+++ text-ui-test/EXPECTED.TXT
@@ -1,4 +1,4 @@
-Hello! I'm Bestie, your personal assistant chatbot.
+Hello! I'm bestie.Bestie, your personal assistant chatbot.
 Let's get ready to have a productive day!
 What can I do for you today :)?
 added: [T][ ] borrow book
